+++ go-service/nexus/test/nexus_test.go
@@ -1,7 +1,8 @@
 // @Author Adrian.Wang 2024/8/30 19:50:00
-package nexus
+package ask
 
 import (
+	"github.com/AdrianWangs/ai-nexus/go-service/nexus/biz/handler/nexus"
 	"github.com/cloudwego/kitex/pkg/klog"
 	"github.com/kr/pretty"
 	"os"
@@ -11,11 +12,11 @@ import (
 func TestGetServicesFromThrift(t *testing.T) {
 
 	/** 设置当前路径为项目根目录 */
-	if err := os.Chdir("../../.."); err != nil {
+	if err := os.Chdir("../"); err != nil {
 		klog.Error("设置当前路径为项目根目录失败")
 		os.Exit(1)
 	}
 
-	res := GetServicesFromThrift()
+	res := nexus.GetServicesFromThrift()
 	pretty.Println(res)
 }
